+++ lib/zego_uikit_prebuilt_video_conference.dart
@@ -3,6 +3,7 @@ library zego_uikit_prebuilt_video_conference;
 export 'package:zego_uikit/zego_uikit.dart';
 
 export 'src/config.dart';
+export 'src/events.dart';
 export 'src/controller.dart';
 export 'src/defines.dart';
 export 'src/video_conference.dart';